indexer: move runner into separate file
